bugfix: coverage is not generated for packages with no code
=== Steps to reproduce ===

1. Create a new directory `foo`.
2. Create `foo/foo.go` with the following content.

    package foo

3. Run `package-coverage .` in `foo`.

Actual result: No coverage is generated, the output is as follows.

    -----------------------------------------------
    |      Branch     |       Dir       |         |
    |   Cov% |  Stmts |   Cov% |  Stmts | Package |
    -----------------------------------------------
    -----------------------------------------------

Expected result: Coverage is generated, the output is as follows.

    -----------------------------------------------
    |      Branch     |       Dir       |         |
    |   Cov% |  Stmts |   Cov% |  Stmts | Package |
    -----------------------------------------------
    | 100.00 |      0 | 100.00 |      0 | foo/    |
    -----------------------------------------------

=== Root cause ===

`go test -coverprofile=profile.cov` does not generate coverage unless there is at least one function in the package.

=== Fix ===

This change creates a file called `fake_code.go` with an empty function. This is similar to how and why we create `fake_test.go`.